Add activite column to apostolat concernes table

+++ database/migrations/2022_04_10_170310_add_activite_column_to_apostolat_concernes_table.php
@@ -14,7 +14,7 @@ class AddActiviteColumnToApostolatConcernesTable extends Migration
     public function up()
     {
         Schema::table('apostolat_concernes', function (Blueprint $table) {
-            $table->unsignedInteger('activite_id');
+            $table->unsignedBigInteger('activite_id');
             $table->foreign('activite_id')->references('id')->on('activites');
         });
     }
